Change unique identifier of a Pet to be its Name
Modified the isSamePet method of Pet such that the check now depends
solely on the names of the pets, instead of using names, species, and
dates of birth or genders. This is to allow a pet to be uniquely
identified by its name, and also enforces a one-to-one relation between
a name and a pet because UniquePetList does not allow duplicates.